use content-type from specified options headers in multipart upload
right now, a different header (in fact, only content-length is specified) is used on PUT calls in multipart upload and the original content-type header is ignored. However, if content-type is not specified, we'll get permission denied from OSS. This commit is to fix in multipart upload of file more than several kbs, the POST call is successful, but ensuing PUT calls fails with 403.